requirements.txt
numpy==1.24.3
opencv-python==4.8.1.78
tensorflow==2.15.0
librosa==0.10.1
scipy==1.11.4
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0 